+++ CharacterControllersDemo/Assets/_SCR/Util/IDamageable.cs
@@ -2,8 +2,8 @@
 
 public interface IDamageable
 {
-    void Damage(int pValue, GameObject pSender, Vector3 pPoint, Vector3 pDirection, Color pColor);
-    void Damage(int pValue, GameObject pSender, Vector3 pPoint, Vector3 pDirection);
+    void Damage(float pValue, GameObject pSender, Vector3 pPoint, Vector3 pDirection, Color pColor);
+    void Damage(float pValue, GameObject pSender, Vector3 pPoint, Vector3 pDirection);
     // void SetStatus(StatusEffects.status pStatus, float pSeconds);
     GameObject GetGameObject();
     IDamageable GetParentDamageable();
